requirements.txt

llama-index
llama-parse
llama-index-embeddings-gemini
llama-index-vector-stores-neo4jvector
llama-index-llms-gemini

google-generativeai
openai
llama-index-embeddings-openai

neo4j

streamlit

python-dotenv
pydantic
pydantic-settings
